use longer workaround timeout on windows

+++ util/all_tests.go
@@ -152,9 +152,17 @@ func sdeOf(cpu, path string, args ...string) (*exec.Cmd, context.CancelFunc) {
 
 	// TODO(CryptoAlg-2154):SDE+ASAN tests will hang without exiting if tests pass for an unknown reason.
 	// Current workaround is to manually cancel the run after 20 minutes and check the output.
-	ctx, cancel := context.WithTimeout(context.Background(), 1200*time.Second)
+	if runtime.GOOS == "linux" {
+		ctx, cancel := context.WithTimeout(context.Background(), 1200*time.Second)
+		return exec.CommandContext(ctx, *sdePath, sdeArgs...), cancel
+	} else if runtime.GOOS == "windows" {
+		// Windows SDE 32 bit also runs into the same issue. The simulation is slower, so we set a timeout
+		// of 30 minutes.
+		ctx, cancel := context.WithTimeout(context.Background(), 1800*time.Second)
+		return exec.CommandContext(ctx, *sdePath, sdeArgs...), cancel
+	}
 
-	return exec.CommandContext(ctx, *sdePath, sdeArgs...), cancel
+	return exec.Command(*sdePath, sdeArgs...), nil
 }
 
 var (
@@ -182,11 +190,12 @@ func runTestOnce(test test, mallocNumToFail int64) (passed bool, err error) {
 		cmd = gdbOf(prog, args...)
 	} else if *useSDE {
 		cmd, cancel = sdeOf(test.cpu, prog, args...)
-		defer cancel()
-
-		cmd.Cancel = func() error {
-			cancelled = true
-			return cmd.Process.Kill()
+		if cancel != nil {
+			defer cancel()
+			cmd.Cancel = func() error {
+				cancelled = true
+				return cmd.Process.Kill()
+			}
 		}
 	} else {
 		cmd = exec.Command(prog, args...)
